requirements.txt
utm~=0.7.0
adafruit_bno055~=1.0.2
board~=1.0
pandas~=1.3.4
gpsd-py3
